Introduced a base class for OASConfigScanClassTest and OASConfigWebInfTest to avoid duplicate @Deployment annotations in a test hierarchy
